+++ crates/macros/src/config_enum/variant.rs
@@ -1,5 +1,7 @@
 use crate::common::FieldSerdeArgs;
-use crate::utils::{extract_comment, extract_common_attrs, format_case, has_attr};
+use crate::utils::{
+    extract_comment, extract_common_attrs, extract_deprecated, format_case, map_option_quote,
+};
 use darling::FromAttributes;
 use proc_macro2::{Ident, TokenStream};
 use quote::quote;
@@ -100,25 +102,17 @@ impl<'l> Variant<'l> {
     }
 
     pub fn get_schema_type(&self) -> TokenStream {
-        let name = self.name.to_string();
-        let value = &self.value;
-        let deprecated = has_attr(&self.attrs, "deprecated");
-
-        let description = if let Some(comment) = extract_comment(&self.attrs) {
-            quote! {
-                Some(#comment.into())
-            }
-        } else {
-            quote! {
-                None
-            }
-        };
+        let name = map_option_quote("name", Some(self.name.to_string()));
+        let description = map_option_quote("description", extract_comment(&self.attrs));
+        let deprecated = map_option_quote("deprecated", extract_deprecated(&self.attrs));
 
         let type_of = if self.args.fallback {
             quote! {
                 SchemaType::string()
             }
         } else {
+            let value = &self.value;
+
             quote! {
                 SchemaType::literal(LiteralValue::String(#value.into()))
             }
@@ -126,10 +120,10 @@ impl<'l> Variant<'l> {
 
         quote! {
             SchemaField {
-                name: Some(#name.into()),
-                description: #description,
                 type_of: #type_of,
-                deprecated: #deprecated,
+                #name
+                #description
+                #deprecated
                 ..Default::default()
             }
         }